configs: sam9x60ek: temporary fix for nand env size
Even if we advertise an u-boot env of 0x40000 bytes (256KB) (see [1]),
use an u-boot env size of 128KB to cope with the existing u-boot code
for the other boards (sama5, sam9x5) and also with the release scripts.

Fixes the following:
Loading Environment from NAND... Can't allocate buffers for environment
*** Warning - malloc() failed, using default environment

The proper fix will be to increase CONFIG_SYS_MALLOC_LEN to be large enough
to comprise both ENV and ENV_REDUNDANT, i.e.

-#define CONFIG_SYS_MALLOC_LEN          (512 * 1024 + 0x1000)
+#define CONFIG_SYS_MALLOC_LEN          (2 * CONFIG_ENV_SIZE + \
+                                        512 * 1024 + 0x1000)

While this was the intention, when doing it I encountered a bad CRC error in
u-boot:
*** Warning - bad CRC, set_default_env using default environment

This is because u-boot is using a 0x40000 ENV_SIZE while the release scripts
use an 0x20000 ENV_SIZE. The fix is to update both u-boot and the release scripts
to use an ENV_SIZE of 0x40000, as described in [1].

Since we are too close to the SAM9X60 IAR release, play safe and be consistent
with what exists already. Proper fix will come for WPS.

[1] http://www.at91.com/linux4sam/pub/Linux4SAM/SambaSubsections//demo_nandflash_map_lnx4sam5x.png

Signed-off-by: Tudor Ambarus <tudor.ambarus@microchip.com>
Reviewed-by: Claudiu Beznea <claudiu.beznea@microchip.com>