ARM: dts: sama7g5ek: disable gmac0
SAMA7G5EK rev. 1 uses KSZ9131 PHY for GMAC0. From time to time the tuple
KSZ9131, GMAC0 doesn't work. This seems to be an issue related to the
interaction b/w these 2 devices, KSZ9131 and GMAC0, as just the enabling of
GMAC0's peripheral clock leads, from time to time, to the blocking of TX
side of the IP and this issue wasn't observed on Evaluation Board where the
KSZ9031 PHY was used. Thus disable the GMAC0 on this rev. 1 board. This
should not affect the GMAC0 functionality on Linux as this issue was
observed only if the U-Boot is in the booting sequence or if the
peripheral clock is enabled in U-Boot in the first seconds after the
hardware reset.

Signed-off-by: Claudiu Beznea <claudiu.beznea@microchip.com>