ARM: mach-at91: Add compile time option to choose proper timer
The sam9x60 with ARM-9 core includes a regular timer and a 64-bit
timer.  This patch adds a compile time option to the Makefile such
that the old timer driver is chosen and compiled as default if none of
timer configuration options are explicitly defined in the board
configs. Similar change has been done for other Soc's belonging to the
AT91 family.

Signed-off-by: Balamanikandan Gunasundar <balamanikandan.gunasundar@microchip.com>