ARM: at91: clock: Update divider and ID handling in at91_periph_clk_enable
Instead of isolating the divider field value that is read from the
register and rewrite it back, use the full register value and ammend
just the bits that we're interested in: AT91_PMC_PCR_EN and
AT91_PMC_PCR_CMD_WRITE. Isolating the divider would make sense if the
function defined a divider parameter and the callers could pass a divider
argument to the method, but it's not the case, so just use the divider
value that is read from the register. The peripheral ID is already written
in a previous register write call, use the peripheral ID that is read from
the register instead of ignoring the read value and overwrite it back.
There are no restrictions in the datasheets with regards to reading the
divider and the ID from PMC and setting them back. The previous stage
bootloader uses a method similar to what I propose, we're safe to modify
at91_periph_clk_enable() for all our SoCs.

Signed-off-by: Tudor Ambarus <tudor.ambarus@microchip.com>