ARM: dts: at91: sama5d2: Align `spi-max-frequency` DT property with Linux
Ensure U-Boot and Linux have the same `spi-max-frequency`. In reality,
any frequency between 55333332 Hz and 82999999 Hz would result in the
`SCBR` field of the `QSPI_SCR` being set by the driver to 2 in the
`atmel_qspi_set_speed` method. This means that the real frequency would
end up somewhere slightly above 55Mhz, enough to be able to read the
JEDEC ID at a proper speed. Before this, the frequency was set to 83Mhz,
which would mean a `SCBR` of 1 and a frequency too high to be able to
properly set up the SPI communication.

Signed-off-by: Sergiu Moga <sergiu.moga@microchip.com>
Reviewed-by: Tudor Ambarus <tudor.ambarus@microchip.com>
Reviewed-by: Eugen Hristev <eugen.hristev@microchip.com>