lds: align u-boot-nodtb with 8 bytes boundary
Newer DTC require that the DTB start address is aligned at 8 bytes.
In the u-boot.bin case, the u-boot-nodtb.bin is concatenated with the
DTB, but there is no alignment/padding to the next 8byte aligned address.
This causes the board to fail booting, because the FDT will claim
that the DTB inside u-boot.bin is not a valid DTB, it will fail with
-FDT_ERR_ALIGNMENT.
To solve this, have the u-boot binary `_end` aligned with 8 bytes.
The objcopy in the Makefile will create the u-boot-nodtb.bin and it has to
be truncated to 8 bytes to correspond to the u-boot.map file which will
have the `_end` aligned to 8 bytes.
The lds files which use devicetrees have been changed to align the `_end`
tag with 8 bytes.

This patch is also a prerequisite to have the possibility to update the
dtc inside u-boot to newer versions (1.6.1+)

Signed-off-by: Eugen Hristev <eugen.hristev@microchip.com>
[manikandan.m@microchip.com: ported to 2024.07]
Signed-off-by: Manikandan Muralidharan <manikandan.m@microchip.com>