include: configs: update boot target order
A bug was found in U-Boot where even if mmc is specified as the boot device, if
there is a QSPI device present with an image then U-Boot - regardless of the
boot script - will boot from QSPI. Create a custom kconfig option that allows
this behaviour to be disabled, i.e. if a user wants to boot from mmc, even
though a QSPI device is set up, they can.

Signed-off-by: Jamie Gibbons <jamie.gibbons@microchip.com>