nand: atmel: Update the smc timings
Update the smc timings in order to probe the nand-flash with ONFI timing
mode 3. This commit is the same that is done for Linux by Mihai.

This change is implemented in at91bootstrap by Bin Li. The comment from his
commit "Some NAND flash (S34ML01G2 and W29N02KVxxAF) do not work properly
in Timing Mode 3, since their maximum tREA time is 4ns longer then normal
NAND flash.  The workaround is to extend the SMC NRD pulse to meet tREA
timing."

This rework is based on the following:
https://bitbucket.microchip.com/projects/LINUX4SAM/repos/at91bootstrap/commits/e2dfd8141d00613a37acee66ef5724f70f34a538#driver%2Fnandflash.c

Suggested-by: Mihai Sain <mihai.sain@microchip.com>
Suggested-by: Li Bin <bin.li@microchip.com>
Signed-off-by: Balamanikandan Gunasundar <balamanikandan.gunasundar@microchip.com>