ARM: dts: at91: sam9x60: Align dts file with linux
Sort the nodes alphabetically and define the pinctrl nodes with its
label. The pinctrl nodes of nand, ebi, macb and few others should be a
child node of pinctrl@fffff400. Without this change those pinctrl nodes are
grouped under an additional 'pinctrl' child node that is not identified by
the pinctrl driver.

Signed-off-by: Balamanikandan Gunasundar <balamanikandan.gunasundar@microchip.com>