ARM: dts: at91: sam9x60_curiosity: define nand ready busy as native pin
Define nand ready/busy as native pin. This setting is meaningless for
sam9x60 SoCs. This is just a workaround that should be fixed asap. This
currently ignores the state of ready busy gpio. Without this change the
ready/busy gpio takes more time to report ready state that makes the nand
flash very slow.

Signed-off-by: Balamanikandan Gunasundar <balamanikandan.gunasundar@microchip.com>