board: microchip: mpfs: make ethernets optional
A given configuration for a board may make either one, or neither of,
the ethernet ports available to U-Boot. The init code will fail if a mac
is not present, so move it to an optional approach.
Node offsets returned by libfdt can contain negative error numbers, so
the variable type should be "int". As things stand, if the ethernet
nodes are not found in the early init callback, the if (node < 0) tests
pass and the code errors out while trying to set the local-mac-address
for a non-existent node.

Signed-off-by: Jamie Gibbons <jamie.gibbons@microchip.com>
Signed-off-by: Conor Dooley <conor.dooley@microchip.com>