ARM: dts: at91: sama7d65: fix sdmmc0 assigned-clock-parents
Fix assigned-clock-parents for sdmmc0 to choose the right parent mck1
instead of syspll.

In SPEC, Table 12.1. Peripheral Identifiers (page 51) states that the
Clock Domain should be MCK1, but it is currently configured with an
incorrect Clock Domain.

This will cause U-Boot to return the error:
mmc@e1204000 - probe failed: -22
and will prevent the SDMMC0 device from being used.

Fixes: 676a9afc39a1 ("ARM: dts: sama7d65: add initial DT for sama7d65 SoC")
Signed-off-by: Varshini Rajendran <varshini.rajendran@microchip.com>
Reviewed-by: Ryan Wanner <ryan.wanner@microchip.com>