net: macb: Add clk init function for SAMA7D65
The SAMA7D65 needs its own clock init function due to the clock
controller limitations. When the clock rate is set in the dts file it is
configured before the parent is enabled which causes unstable behavior
for this SoC. By having the driver set the clock rate this ensures the
parent clock is enabled before the clock is fully configured.

The rate of the tx_clk is set by the driver during initialization
instead of fixed in the device tree.

Signed-off-by: Ryan Wanner <Ryan.Wanner@microchip.com>